Add sidebar utils
The categories function tries to gather all 3D view sidebar categories bpy looking at the data of every subclass of bpy.types.Panel.
It seems to work pretty well, though some of the categories that are returned don't actually show in the 3D view, so there's still work to do.