Add functionality to the Open Tab operator
It's not ideal, but it's better than nothing.
Because I can't figure out how to change which tab we're on, and can't seem to pin panels using code, I just change the bl_category of each panel, and re-register them.
The original category is stored inside the panels themselves.
The order of if statements and the use of elif is deliberate; this way if you choose SIMPLE TABS, it'll move the currently moved panels back, and do nothing else.