Add check for "Item" tab
This tab contains a panel that's probably defined in C rather than Python, which can not be unregistered. Therefore, it can't be renamed or moved or anything. So I'm just gonna exclude it from the list and pretend it's not there, which will put it at the top of the sidebar.