Add TabProps property group
This stores the original name and new name for each sidebar tab.
I've made it so that setting the new name to empty will reset it to the original name, because I think it's convenient.
Originally I had planned to make it hide the tab entirely, but I've since found that that would be a logistical nightmare, and not that useful anyway.
Oh and the order of registration for the props submodule is changed slightly, to make sure that interdependency doesn't cause any trouble.