Add Update operator
This one is a bit complicated.
First, get the panels, and filter out ones that aren't in any of the tabs in our list.
Then, set the bl_category to the original_category on any panels that have it, and delete the original_category from them.
Next, store the bl_category in original_category for each panel, and set the bl_category to the new category according to our list.
Then, sort the panels into lists per tab, which in turn are sorted according to our list.
And finally, go through the sorted panels and unregister / register them.
Sounds like a lot, but it's instant, at least on my setup.