Add new UI to panel
I'm making the UIList have the length of the list of tabs, because I think it looks nice and works well.
The buttons are on the side, as with almost all UILists in Blender.
I made them 1.5 times as big (at reduced width so they remain square) because I think it looks better, and is easier to click on.
I like the icons I chose for the buttons, but I can change them if you have better ones in mind.