Remove bl_context check from panel check util
The problem is, this tool should work consistency across all modes, without having to refresh and update every time you switch to edit mode and such. So, we need to get as many panels as possible, which means this particular check needs to be removed and replaced with something better.