Take bl_parent_id into account in sidebar utils
In the check / tab function, if a panel has a parent panel ID, and that parent panel is registered, then call the check / tab function, with the parent panel.
Also I'm calling tab from check instead of reading bl_category directly there.
And I changed the elifs to ifs in the tab function, because it was necessary for the nested if statement to work properly.