Use timers submodule instead of temp code
This was only here because I had to put it *somewhere*, so now I'm removing it and using the timers submodule I made instead.
Btw if you're wondering why I'm not calling the refresh operator, it's because it removes categories that are no longer around, which is destructive.
Maybe in the future we can come up with something non-destructive I guess. Anyway it's fine for the time being.