Update sidebar utility functions
Where do I start...
I've moved some pieces of code around.
Instead of a hardcoded check for the Item tab, it's now checked against a list from preferences.
I now filter out panels that aren't registered (I wasn't doing this before because it wasn't working, but that's because I was doing it wrong, and now I'm doing it right).
Not doing the poll check on BoxCutter panels made it so that the startup timer worked on them, but it also caused several panels to show up that definitely shouldn't have, so I'm reverting that to just being blender native panels again.
It's still something to think about I guess.
Anyway..
I've improved the panels() function, it now gathers subclasses recursively, so that fixes the situation with Zen UV.
Also I improved the update function so it gives a more clear explanation about *why* something failed, and doesn't try to register a class that wasn't registered in the first place.