Bump version to 1.1.7
This update gives you a popup if a panel has no module (this should never happen, but it for one customer), and always re-registers subpanels if their parent passes the check (this needs more testing to be safe).
Also I'm adjusting bl_info a little, removing the empty wiki_url since it's useless, and adding a trailing comma.