Fix panel recursion
I've become a better programmer since the time I wrote this code.
Notice that the recursive function never adds class A to the panels list.
This means that, because we call the function on subclasses of class B, instead of on class B itself, class B never gets added to the panels list.
What that means is, that every other generation of subclasses gets skipped, which makes it so most addons that use a base panel class don't work here.
It was a simple fix, just call the recursive function on class B directly, instead of its subclasses.