Filter problem panel instead of Item tab
The reason the Item category was excluded by default is because of this pesky panel that refuses to (un)register, so I'll just filter for it specifically, instead of the whole category.
This won't make a difference to users, unless they have an addon that adds something to the Item category; they'll be able to move it now.