Make move and refresh buttons smarter about index
When you refresh the list, it'll now select the first item, if none was selected before.
When you press a move button with no item selected (this should never happen because of the refresh change, but just in case), it'll just do nothing (previously it would select the first item).
(Unrelated but, when you set the active material index on an object to -1 and then try to use the move buttons, it'll crash to desktop).