Bump version to 1.2.0
This version makes it so SIMPLE TABS no longer forgets tabs; instead you can remove them manually if need be.
The "Item" tab is no longer a forced part of the exclude list, but the troublesome panel inside it is hardcoded to be ignored.
The preferences button now opens the SIMPLE TABS preferences instead of saving them, because saving is now done automatically when you make changes.
The UI is smaller to fit the extra buttons more comfortably.
The sidebar no longer automatically opens on refresh/update, but this can be turned on in the preferences.