Bump version to 1.2.3
This update makes it so the Item and Tool tabs are always excluded.
A user reported that renaming the Tool tab messes up various UI elements in Blender, so I think the best solution is to simply not mess with that category.
I also added the Item tab, because it has issues too; the panel you'll find there doesn't cooperate.